fix(containers)!: race conditions between multiple app instances
When running tests in Wing, an isolated environment is created for every test. This means we must create a the container workload using a unique name, or otherwise the containers will mix up with each other.

BREAKING CHANGE: the `getPublicUrl()` and `getInternalUrl()` APIs were replaced with properties (`publicUrl` and `internalUrl`).