feat(dynamodb)!: api tweaks
## `connection`

`table.connection` now includes `clientConfig` instead of just `endpoint`, so it can be used to create an AWS SDK client without any hacks.

I also changed it from a method to a property, so it can be lifted across inflight.

## `tableName`

Added a `tableName` property with the name of the table.

Found a bug in `tf-aws` where `name` was not respected.

Tested on AWS.