chore: use indexing syntax (#204)
Cosmetic refactorings to our winglibs to use the new indexing syntax where applicable. This is mostly done as an effort to test some usage of the syntax in the real world to catch if there are any bugs. One bug was found and since fixed [here](https://github.com/winglang/wing/pull/6277).